Move each page component in App.tsx to new pages directory
-Move each page component in App.tsx to new pages directory
-create type CollectionFromSearchPayload to fix payload error
-change ProductDetailPage Wrapper onEdit field to use productUpdate instead of product
-remove unused imports in App.tsx
