+++ src/pages/ProductDetailPage.tsx
@@ -1,0 +1,259 @@
+import React, { useEffect, useState } from 'react'
+import { useNavigate, useParams, useSearchParams } from 'react-router-dom'
+import { Button } from '@/components/ui/button'
+import { ProductDetail } from '@/components/ProductDetail'
+import { APIService } from '@/lib/api'
+import { Collection, CollectionCreateInput, Discussion, Product, ProductUpdate, Rating, UserAccount, UserData } from '@/lib/types'
+import { toast } from 'sonner'
+
+function ProductDetailPage({
+    products,
+    ratings,
+    discussions,
+    user,
+    userAccount,
+    userCollections,
+    onRate,
+    onDiscuss,
+    onAddTag,
+    onAddToCollection,
+    onRemoveFromCollection,
+    onCreateCollection,
+    onDelete,
+    onEdit,
+    onToggleBan,
+    onEditDiscussion,
+    onDeleteDiscussion,
+    onToggleBlockDiscussion,
+    allTags,
+    allProductTypes = [],
+}: {
+    products: Product[]
+    ratings: Rating[]
+    discussions: Discussion[]
+    user: UserData | null
+    userAccount: UserAccount | null
+    userCollections: Collection[]
+    onRate: (productId: string, rating: number) => void
+    onDiscuss: (productId: string, content: string, parentId?: string) => void
+    onAddTag: (productId: string, tag: string, productObj?: Product) => void
+    onAddToCollection: (collectionSlug: string) => Promise<void>
+    onRemoveFromCollection: (collectionSlug: string) => Promise<void>
+    onCreateCollection: (data: CollectionCreateInput) => void
+    onDelete: (productId: string) => void
+    onEdit: (product: ProductUpdate) => void
+    onToggleBan: (product: Product, reason?: string) => void
+    onEditDiscussion: (id: string, content: string) => Promise<void> | void
+    onDeleteDiscussion: (id: string) => Promise<void> | void
+    onToggleBlockDiscussion: (id: string, block: boolean) => Promise<void> | void
+    allTags: string[]
+    allProductTypes?: string[]
+}) {
+    const { slug: productSlug } = useParams()
+    const [searchParams] = useSearchParams()
+    const autoOpenEdit = searchParams.get('edit') === '1'
+    const autoOpenOwnershipRequest = searchParams.get('requestEdit') === '1'
+    const navigate = useNavigate()
+    const [product, setProduct] = useState<Product | null>(null)
+    const [loading, setLoading] = useState(true)
+
+    useEffect(() => {
+        const fetchProduct = async () => {
+            if (!productSlug) {
+                setLoading(false)
+                return
+            }
+
+            try {
+                // First check if product is already in the products array (from home page navigation)
+                const cachedProduct = products.find(p => p.slug === productSlug)
+                if (cachedProduct) {
+                    setProduct({ ...cachedProduct, slug: cachedProduct.slug ?? productSlug })
+                    setLoading(false)
+                    return
+                }
+
+                // Otherwise fetch just this product
+                const fetchedProduct = await APIService.getProductBySlug(productSlug)
+                setProduct(fetchedProduct ? { ...fetchedProduct, slug: fetchedProduct.slug ?? productSlug } : null)
+            } catch (error) {
+                console.error('Failed to fetch product:', error)
+                setProduct(null)
+            } finally {
+                setLoading(false)
+            }
+        }
+
+        fetchProduct()
+    }, [productSlug, products])
+
+    if (loading) {
+        return (
+            <div className="text-center py-12">
+                <p className="text-lg text-muted-foreground">Loading...</p>
+            </div>
+        )
+    }
+
+    if (!product) {
+        return (
+            <div className="text-center py-12">
+                <p className="text-lg text-muted-foreground">Product not found</p>
+                <Button variant="outline" onClick={() => navigate('/')} className="mt-4">
+                    Back to Products
+                </Button>
+            </div>
+        )
+    }
+
+    return (
+        <ProductDetail
+            product={product}
+            ratings={ratings}
+            discussions={discussions}
+            user={user}
+            userAccount={userAccount}
+            userCollections={userCollections}
+            onBack={() => navigate('/')}
+            onRate={(rating) => onRate(product.id, rating)}
+            onDiscuss={(content, parentId) => onDiscuss(product.id, content, parentId)}
+            onAddTag={(tag) => onAddTag(product.slug ?? product.id, tag, product)}
+            onAddToCollection={onAddToCollection}
+            onRemoveFromCollection={onRemoveFromCollection}
+            onCreateCollection={onCreateCollection}
+            allTags={allTags}
+            allProductTypes={allProductTypes}
+            onDelete={onDelete}
+            onEdit={onEdit}
+            onToggleBan={onToggleBan}
+            onEditDiscussion={onEditDiscussion}
+            onDeleteDiscussion={onDeleteDiscussion}
+            onToggleBlockDiscussion={onToggleBlockDiscussion}
+            autoOpenEdit={autoOpenEdit}
+            autoOpenOwnershipRequest={autoOpenOwnershipRequest}
+        />
+    )
+}
+
+export function ProductDetailPageWrapper({
+    products,
+    ratings,
+    discussions,
+    user,
+    userAccount,
+    userCollections,
+    onRate,
+    onDiscuss,
+    onAddTag,
+    onCollectionsChange,
+    onCreateCollection,
+    onDelete,
+    onEdit,
+    onToggleBan,
+    onEditDiscussion,
+    onDeleteDiscussion,
+    onToggleBlockDiscussion,
+    allTags,
+    allProductTypes = [],
+}: {
+    products: Product[]
+    ratings: Rating[]
+    discussions: Discussion[]
+    user: UserData | null
+    userAccount: UserAccount | null
+    userCollections: Collection[]
+    onRate: (productId: string, rating: number) => void
+    onDiscuss: (productId: string, content: string, parentId?: string) => void
+    onAddTag: (productId: string, tag: string, productObj?: Product) => void
+    onCollectionsChange: (collections: Collection[] | ((current: Collection[]) => Collection[])) => void
+    onCreateCollection: (data: CollectionCreateInput) => void
+    onDelete: (productId: string) => void
+    onEdit: (product: ProductUpdate) => void
+    onToggleBan: (product: Product, reason?: string) => void
+    onEditDiscussion: (id: string, content: string) => Promise<void> | void
+    onDeleteDiscussion: (id: string) => Promise<void> | void
+    onToggleBlockDiscussion: (id: string, block: boolean) => Promise<void> | void
+    allTags: string[]
+    allProductTypes?: string[]
+}) {
+    const { slug } = useParams()
+    const [searchParams] = useSearchParams()
+    void searchParams
+
+    const [localRatings, setLocalRatings] = useState<Rating[]>(ratings)
+    const [localDiscussions, setLocalDiscussions] = useState<Discussion[]>(discussions)
+
+    // Fetch ratings and discussions if not already loaded
+    useEffect(() => {
+        const fetchData = async () => {
+            if (ratings.length === 0) {
+                try {
+                    const [allRatings, allDiscussions] = await Promise.all([
+                        APIService.getAllRatings(),
+                        APIService.getAllDiscussions(),
+                    ])
+                    setLocalRatings(allRatings)
+                    setLocalDiscussions(allDiscussions)
+                } catch (error) {
+                    console.warn('[ProductDetailPageWrapper] Failed to fetch ratings/discussions:', error)
+                }
+            }
+        }
+        fetchData()
+    }, [ratings.length])
+
+    // Keep local state in sync when parent updates
+    useEffect(() => {
+        if (ratings.length > 0) setLocalRatings(ratings)
+    }, [ratings])
+
+    useEffect(() => {
+        if (discussions.length > 0) setLocalDiscussions(discussions)
+    }, [discussions])
+
+    const handleAddToCollection = async (collectionSlug: string) => {
+        if (!slug) return
+
+        if (!collectionSlug) return
+
+        const updated = await APIService.addProductToCollection(collectionSlug, slug)
+        if (updated) {
+            onCollectionsChange((current) => current.map((c) => ((c.slug || c.id) === collectionSlug ? updated : c)))
+            toast.success('Added to collection')
+        }
+    }
+
+    const handleRemoveFromCollection = async (collectionSlug: string) => {
+        if (!slug) return
+        const updated = await APIService.removeProductFromCollection(collectionSlug, slug)
+        if (updated) {
+            onCollectionsChange((current) => current.map((c) => ((c.slug || c.id) === collectionSlug ? updated : c)))
+            toast.success('Removed from collection')
+        }
+    }
+
+    return (
+        <ProductDetailPage
+            products={products}
+            ratings={localRatings}
+            discussions={localDiscussions}
+            user={user}
+            userAccount={userAccount}
+            userCollections={userCollections}
+            onRate={onRate}
+            onDiscuss={onDiscuss}
+            onAddTag={onAddTag}
+            onAddToCollection={handleAddToCollection}
+            onRemoveFromCollection={handleRemoveFromCollection}
+            onCreateCollection={onCreateCollection}
+            onDelete={onDelete}
+            onEdit={onEdit}
+            onToggleBan={onToggleBan}
+            onEditDiscussion={onEditDiscussion}
+            onDeleteDiscussion={onDeleteDiscussion}
+            onToggleBlockDiscussion={onToggleBlockDiscussion}
+            allTags={allTags}
+            allProductTypes={allProductTypes}
+        />
+    )
+}
